Updated image handling structure
