+++ Controllers/EmployeeController.cs
@@ -1,6 +1,4 @@
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using OOL_API.Models;
 using OOL_API.Models.DataTransfer;
 using OOL_API.Services;
 
@@ -14,47 +12,9 @@ namespace OOL_API.Controllers
     {
         private readonly CurrentUserInfo _currentUserInfo;
 
-        private readonly IPictureStorage<Employee, string> _pictureStorage;
-
-        public EmployeeController(
-            CurrentUserInfo currentUserInfo,
-            IPictureStorage<Employee, string> pictureStorage
-        )
+        public EmployeeController(CurrentUserInfo currentUserInfo)
         {
             _currentUserInfo = currentUserInfo;
-            _pictureStorage = pictureStorage;
-        }
-
-        [HttpGet]
-        [Route("picture")]
-        public IActionResult GetPicture()
-        {
-            var employee = _currentUserInfo.GetCurrentEmployee();
-
-            if (employee == null) return Unauthorized();
-
-            var content = _pictureStorage.GetPicture(employee.UserId);
-
-            if (content == null) return NotFound();
-
-            return File(content, "image/jpeg");
-        }
-
-        [HttpPost]
-        [Route("upload-image")]
-        public IActionResult Upload([FromForm] IFormFile file)
-        {
-            var employee = _currentUserInfo.GetCurrentEmployee();
-
-            if (employee == null) return Unauthorized();
-
-            if (file.ContentType != "image/jpeg") return BadRequest();
-
-            using var stream = file.OpenReadStream();
-
-            _pictureStorage.PostPicture(stream, employee);
-
-            return CreatedAtAction("GetPicture", "");
         }
 
         [HttpGet]
